completed user model and handler testing
